requirements.txt
brisque
pillow
numpy
rich
